refactor(domains): rename helpers to service, stakingUtils to stakingCommonService
