+++ src/renderer/features/staking-bond-nominate/model/form-model.ts
@@ -19,7 +19,7 @@ import {
 } from '@/shared/lib/utils';
 import { createComplexTxStore, createSignatoriesStore, createTxValidationStore } from '@/shared/transactions';
 import { type AnyAccount, accountService, accounts } from '@/domains/network';
-import { stakingUtils } from '@/domains/staking';
+import { stakingCommonService } from '@/domains/staking';
 import { balanceModel, balanceUtils } from '@/entities/balance';
 import { networkModel } from '@/entities/network';
 import { transactionBuilder } from '@/entities/transaction';
@@ -331,7 +331,7 @@ const $reusableLock = combine({ balance: $initiatorBalance, available: $availabl
     return null;
   }
 
-  const reusableLock = stakingUtils.reusableLockBN(balance);
+  const reusableLock = stakingCommonService.reusableLockBN(balance);
   return BN.min(available, reusableLock);
 });
